requirements-training.txt
torch>=2.0.0  # PyTorch for deep learning
transformers>=4.36.0  # Hugging Face Transformers
datasets>=2.14.0  # Hugging Face Datasets
accelerate>=0.25.0  # For efficient training
bitsandbytes>=0.41.0  # For 4-bit quantization
peft>=0.7.0  # For LoRA fine-tuning
trl>=0.7.4
wandb>=0.15.12  # For experiment tracking
deepspeed>=0.12.0  # For distributed training (optional)
faiss-cpu>=1.7.4  # for vector search
requests>=2.31.0  # for API calls
python-dotenv>=1.0.0  # For environment variable management
schedule>=1.2.0  # for task scheduling
pydantic>=2.0.0  # for data validation
rich>=13.7.0  # for beautiful terminal output
colorama>=0.4.6  # for colored terminal output
sentence-transformers>=2.2.0  # for RAG embeddings
scipy>=1.10.0
numpy<2.0.0  # Downgrade to 1.x for compatibility
tqdm>=4.65.0
rouge-score>=0.1.2  # for ROUGE evaluation metrics
gradio>=4.0.0  # for web interface
llama-index>=0.9.0  # for RAG system
chromadb>=0.4.22  # for vector store
SpeechRecognition>=3.10.0  # for voice control
googletrans>=3.1.0a0  # for translation
langdetect>=1.0.9  # for language detection
moviepy>=1.0.3  # for video editing
elevenlabs>=0.2.27  # for voice generation
pandas>=2.0.0  # for data analysis
plotly>=5.18.0  # for interactive charts
instagram-private-api>=1.6.0  # for Instagram integration
safetensors>=0.4.0  # for efficient model loading
protobuf>=4.25.1  # for model serialization 